+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-All notable changes to the "usage-heatmap" extension will be documented in this file.
+All notable changes to the "change-heatmap" extension will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2024-01-01
 
 ### Added
-- Initial release of Usage Heatmap VS Code extension
+- Initial release of Change Heatmap VS Code extension
 - Display current cursor line number in the status bar
 - Faint grey color styling for non-intrusive experience
 - Real-time updates as cursor moves through files